client, keys: move mocks into package mock
